+++ src/main/java/te/trueEcho/domain/friend/entity/FriendStatus.java
@@ -1,4 +1,4 @@
-package te.trueEcho.domain.friend;
+package te.trueEcho.domain.friend.entity;
 
 public enum FriendStatus {
     FRIEND, NOT_FRIEND, REQUEST
