+++ src/main/java/te/trueEcho/domain/notification/entity/LikeNoti.java
@@ -1,10 +1,10 @@
-package te.trueEcho.domain.notification;
+package te.trueEcho.domain.notification.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.NoArgsConstructor;
-import te.trueEcho.domain.post.Like;
+import te.trueEcho.domain.post.entity.Like;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
