+++ src/main/java/te/trueEcho/domain/notification/entity/Notification.java
@@ -1,4 +1,4 @@
-package te.trueEcho.domain.notification;
+package te.trueEcho.domain.notification.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
