+++ src/main/java/te/trueEcho/domain/notification/entity/ServiceNoti.java
@@ -1,4 +1,4 @@
-package te.trueEcho.domain.notification;
+package te.trueEcho.domain.notification.entity;
 
 
 import jakarta.persistence.*;
