+++ src/main/java/te/trueEcho/domain/post/entity/Pin.java
@@ -1,4 +1,4 @@
-package te.trueEcho.domain.post;
+package te.trueEcho.domain.post.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
@@ -7,7 +7,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-import te.trueEcho.domain.user.User;
+import te.trueEcho.domain.user.entity.User;
 import java.time.LocalDateTime;
 
 @Getter
