+++ src/main/java/te/trueEcho/domain/post/entity/PostStatus.java
@@ -1,4 +1,4 @@
-package te.trueEcho.domain.post;
+package te.trueEcho.domain.post.entity;
 
 public enum PostStatus {
     FREETIME, ONTIME, LATE
