+++ src/main/java/te/trueEcho/domain/user/entity/Block.java
@@ -1,38 +1,35 @@
-package te.trueEcho.domain.friend;
+package te.trueEcho.domain.user.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import te.trueEcho.domain.user.User;
 
 @Getter
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "friends")
-public class Friend {
+@Table(name = "blocks")
+public class Block {
 
     @Id
-    @Column(name = "friend_id")
+    @Column(name = "block_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-
-    @Column(name = "friend_status")
-    private FriendStatus friendStatus;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id")
     private User user;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "target_user_id")
-    private User targetUser;
+    @JoinColumn(name = "block_user_id")
+    private User blockUser;
 
+    @Column(name = "block_level")
+    private int blockLevel;
     @Builder
-    public Friend(FriendStatus friendStatus, User user, User targetUser) {
-        this.friendStatus = friendStatus;
+    public Block(User user, User blockUser) {
         this.user = user;
-        this.targetUser = targetUser;
+        this.blockUser = blockUser;
     }
 }
