+++ src/main/java/te/trueEcho/domain/user/entity/Gender.java
@@ -1,4 +1,4 @@
-package te.trueEcho.domain.user;
+package te.trueEcho.domain.user.entity;
 
 public enum Gender {
     // 남성, 여성, 밝히고싶지않음
